make_bouding_box/pyproject.toml
[project]
name = "make-bouding-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub>=0.3.13",
    "numpy>=2.3.2",
    "opencv-contrib-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
